requirements.txt
pandas>=1.5.0
numpy>=1.21.0
duckdb>=0.9.0
pyarrow>=12.0.0
apache-airflow>=2.7.0
